Implemented "Extended Lineal Base Address" of HEX files.

+++ Arduimmer/HexRecord.cs
@@ -38,7 +38,7 @@ namespace Arduimmer
 			get { return ':'; }
 		}
 
-		public ushort Address {
+		public uint Address {
 			get;
 			set;
 		}
